requirements.txt
alembic==1.4.0
amqp==2.5.2
bcrypt==3.1.7
beautifulsoup4==4.8.2
billiard==3.6.2.0
blinker==1.4
bs4==0.0.1
celery==4.4.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cssmin==0.2.0
defusedxml==0.6.0
Flask==1.1.1
Flask-Admin==1.5.5
Flask-Assets==2.0
Flask-Bcrypt==0.7.1
Flask-Celery-Helper==1.1.0
Flask-CKEditor==0.4.3
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-OAuth==0.12
Flask-OpenID==1.2.5
Flask-PageDown==0.2.2
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
gunicorn==20.0.4
httplib2==0.17.0
idna==2.8
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsmin==2.2.2
kombu==4.6.7
Mako==1.1.1
Markdown==3.2.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
oauth2==1.9.0.post1
pycparser==2.19
PyMySQL==0.9.3
python-dateutil==2.8.1
python-editor==1.0.4
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
six==1.14.0
soupsieve==1.9.5
SQLAlchemy==1.3.13
urllib3==1.25.8
vine==1.3.0
webassets==2.0
Werkzeug==0.16.1
wtf==0.1
WTForms==2.2.1
zipp==3.0.0
